requirements.txt
asttokens==2.4.1
colorama==0.4.6
comm==0.2.1
debugpy==1.8.0
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
importlib-metadata==7.0.1
ipykernel==6.29.0
ipython==8.18.1
jedi==0.19.1
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.7.1
matplotlib-inline==0.1.6
nest-asyncio==1.6.0
numpy==1.26.3
packaging==23.2
pandas==2.2.0
parso==0.8.3
platformdirs==4.1.0
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
py4j==0.10.9.7
Pygments==2.17.2
pyspark==3.5.0
python-dateutil==2.8.2
pytz==2023.4
pywin32==306
pyzmq==25.1.2
scikit-surprise==1.1.3
scipy==1.12.0
six==1.16.0
stack-data==0.6.3
tornado==6.4
traitlets==5.14.1
typing_extensions==4.9.0
tzdata==2023.4
wcwidth==0.2.13
zipp==3.17.0
